Rename serde helpers to be consistent
