+++ src/serde_utils/mod.rs
@@ -2,8 +2,8 @@ pub mod hex_fixed_vec;
 pub mod hex_prog_var_list;
 pub mod hex_var_list;
 pub mod list_of_hex_fixed_vec;
-pub mod list_of_hex_prog_fixed_vec;
-pub mod list_of_hex_prog_var_list;
 pub mod list_of_hex_var_list;
+pub mod prog_list_of_hex_fixed_vec;
+pub mod prog_list_of_hex_prog_var_list;
 pub mod quoted_u64_fixed_vec;
 pub mod quoted_u64_var_list;
